Remove backend – use static json files only
This removes the Cloud Function, Firestore and Authentication and
Kramster is now entirely a static frontend app.

This means no reports or stats are generated anymore and it is not
possible to know which questions are the hardest based on stats.

On the positive side, the app is simpler, faster, and not reliant on
Google services to run (although still using Firebase Hosting and some
Google Fonts for now).

It will also be easier for people to contribute with new exams through
pull requests. There are also fewer dependencies and less maintenance
will be needed in the future.